1. copy exe and dll to build/bin 2. copy .a to build/lib 3. copy headers to build/include

+++ build.bat
@@ -1,5 +1,5 @@
 
-rmdir /s /q build
+if exist build rmdir /s /q build
 mkdir build
 cd build
 cmake -DPROTOBUF_ROOT=d:/work/protobuf-2.4.1 -DCMAKE_BUILD_TYPE=Release -G "MinGW Makefiles" ..
